+++ contest_3/c_first_header/sum.h
@@ -1,6 +1,8 @@
 #ifndef H_SUM
 #define H_SUM
 
-int Sum(int x, int y);
+#include <cstdint>
+
+int64_t Sum(int x, int y);
 
 #endif
